+++ packages/camera/camera_platform_interface/test/method_channel/method_channel_camera_test.dart
@@ -8,6 +8,7 @@ import 'dart:math';
 import 'package:async/async.dart';
 import 'package:camera_platform_interface/camera_platform_interface.dart';
 import 'package:camera_platform_interface/src/method_channel/method_channel_camera.dart';
+import 'package:camera_platform_interface/src/types/focus_mode.dart';
 import 'package:camera_platform_interface/src/utils/utils.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
@@ -25,7 +26,10 @@ void main() {
         MethodChannelMock cameraMockChannel = MethodChannelMock(
             channelName: 'plugins.flutter.io/camera',
             methods: {
-              'create': {'cameraId': 1}
+              'create': {
+                'cameraId': 1,
+                'imageFormatGroup': 'unknown',
+              }
             });
         final camera = MethodChannelCamera();
 
@@ -108,7 +112,10 @@ void main() {
         MethodChannelMock cameraMockChannel = MethodChannelMock(
             channelName: 'plugins.flutter.io/camera',
             methods: {
-              'create': {'cameraId': 1},
+              'create': {
+                'cameraId': 1,
+                'imageFormatGroup': 'unknown',
+              },
               'initialize': null
             });
         final camera = MethodChannelCamera();
@@ -125,6 +132,8 @@ void main() {
           1080,
           ExposureMode.auto,
           true,
+          FocusMode.auto,
+          true,
         ));
         await initializeFuture;
 
@@ -136,6 +145,7 @@ void main() {
             'initialize',
             arguments: {
               'cameraId': 1,
+              'imageFormatGroup': 'unknown',
             },
           ),
         ]);
@@ -162,6 +172,8 @@ void main() {
           1920,
           1080,
           ExposureMode.auto,
+          true,
+          FocusMode.auto,
           true,
         ));
         await initializeFuture;
@@ -205,6 +217,8 @@ void main() {
           1080,
           ExposureMode.auto,
           true,
+          FocusMode.auto,
+          true,
         ));
         await initializeFuture;
       });
@@ -221,6 +235,8 @@ void main() {
           3840,
           2160,
           ExposureMode.auto,
+          true,
+          FocusMode.auto,
           true,
         );
         await camera.handleMethodCall(
@@ -332,6 +348,8 @@ void main() {
             1920,
             1080,
             ExposureMode.auto,
+            true,
+            FocusMode.auto,
             true,
           ),
         );
@@ -668,6 +686,54 @@ void main() {
           isMethodCall('setExposureOffset', arguments: {
             'cameraId': cameraId,
             'offset': 0.5,
+          }),
+        ]);
+      });
+
+      test('Should set the focus mode', () async {
+        // Arrange
+        MethodChannelMock channel = MethodChannelMock(
+          channelName: 'plugins.flutter.io/camera',
+          methods: {'setFocusMode': null},
+        );
+
+        // Act
+        await camera.setFocusMode(cameraId, FocusMode.auto);
+        await camera.setFocusMode(cameraId, FocusMode.locked);
+
+        // Assert
+        expect(channel.log, <Matcher>[
+          isMethodCall('setFocusMode',
+              arguments: {'cameraId': cameraId, 'mode': 'auto'}),
+          isMethodCall('setFocusMode',
+              arguments: {'cameraId': cameraId, 'mode': 'locked'}),
+        ]);
+      });
+
+      test('Should set the exposure point', () async {
+        // Arrange
+        MethodChannelMock channel = MethodChannelMock(
+          channelName: 'plugins.flutter.io/camera',
+          methods: {'setFocusPoint': null},
+        );
+
+        // Act
+        await camera.setFocusPoint(cameraId, Point<double>(0.5, 0.5));
+        await camera.setFocusPoint(cameraId, null);
+
+        // Assert
+        expect(channel.log, <Matcher>[
+          isMethodCall('setFocusPoint', arguments: {
+            'cameraId': cameraId,
+            'x': 0.5,
+            'y': 0.5,
+            'reset': false
+          }),
+          isMethodCall('setFocusPoint', arguments: {
+            'cameraId': cameraId,
+            'x': null,
+            'y': null,
+            'reset': true
           }),
         ]);
       });