+++ packages/camera/camera_platform_interface/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.4.0
+
+- Added interface methods to support auto focus.
+
+## 1.3.0
+
+- Introduces an option to set the image format when initializing.
+
 ## 1.2.0
 
 - Added interface to support automatic exposure.